libutil: cf: pass copy to update_file() in cf_update_glob()
Problem: cf_update_glob() first updates a new, empty cf_t object
before then applying updates to the caller's cf object. This means
that any existing configuration in the passed cf object is lost on
transition to update_file().

Pass a copy of the cf object instead of a new cf object to
update_file() in cf_update_glob() so that any existing configuration
in the cf argument is preserved and available in update_file().