imp: configure test IMP without config file security
Problem: The build-tree version of the IMP is used in sharness tests,
where it is convenient to create temporary config files directly in
the sharness "trash" directory. However, this use case may cause the
file and directory security checks in cf_update_* to fail.

Unless FLUX_TEST_IMP_PATH_PARANOIA is set in the environment, disable
file and directory security checks for IMP and flux-security context
in the test IMP only.

This should allow all existing IMP tests to pass, with an option to
disable this behavior for IMP security checks.