testsuite: verify cf class path security checks
Problem: The cf class security checks are only enabled when the
effective userid does not match the current userid, or the effective
userid is 0, so 'paranoid' mode is difficult if not impossible
to test in the unit test framework.

Instead, use the `t/src/cf` test program with sudo to simulate
tests in a real-world scenario (where the IMP is installed setuid).
If the SUDO prereq is met, run a series of tests that ensure common
file and directory security checks are being completed by the
cf_update_glob(3) call.