build: add 'make deb' target to build test debian package
Problem: There is no provided method to package a build directory
for testing.

Add a set of debian packaging metadata under the debian directory,
along with a script in scripts/debbuild.sh to package the current
source.

For convenience, also add a 'make deb' target to the top level
Makefile.am which will use the new script to build a debian package