imp: add a simple and safe implementation of popen(3)
Problem: The IMP will need popen-like functionality to fork an optional
helper process to gather input, but popen(3) unnecessarily calls out
to the shell which could be insecure.

Add a safer version of popen(3) in imp/exec/safe_popen.[ch].

This version splits the command argument into tokens separated by
whitespace and passes that directly to execvp(2) instead of passing to
`sh -c`.