testsuite: t2000-imp-exec.t: fix expected error tests
Problem: The initial few expected failure tests in t2000-imp-exec.t
"succeed" because the current user is not configured to use the IMP,
not because of bad input or an invalid commandline.

Add a config file that allows the user executing the testsuite to use
the IMP so that the tests fail for the expected reason.