imp: exec: issue error if pam-support used without --enable-pam
Problem: The IMP silently ignores `exec.pam-support = true` in the
config file when flux-security was not built with --enable-pam. This
can lead to confusion or worse if an admin erroneously thinks PAM
support is working when it is not.

Generate a fatal error if `exec.pam-support` is enabled when the IMP
is not built with --enable-pam.