ci: update ci test images
Problem: LTS for the bionic base image for CI testing has expired,
and there is no longer a need to support RHEL 7, therefore it is
probably time to update the images used in the CI testing for the
flux-security project.

Replace bionic with the Debian bookworm image (Debian 12.0), drop
the el7 based tests, and replace fedora35 with fedora 38.