imp: discover current cgroup information at IMP startup
Problem: To comply with RFC 15, the IMP needs to know attributes of
its current cgroup, but this information is not currently available.

Add a cgroup_info structure to the imp_state object, and populate it
at IMP startup.

This moves and modifies some internal code from pidinfo.c. In the
future, this duplicate code in pidinfo.c will be removed.