imp: support SIGUSR1 as surrogate for SIGKILL in imp exec
Problem: RFC 15 specifies that the IMP should handle SIGUSR1 as a
surrogate for SIGKILL, but this is not currently supported.

Add cgroup_kill() which will send a signal to all member's of the
IMP's cgroup (besides the IMP's own pid).

Add SIGUSR1 to the set of signals handled by the signal forwarding
implementation of the IMP. In the case of SIGUSR1, deliver SIGKILL
to all members of the current cgroup if use_cgroup_kill is true,
otherwise fallback to pid_kill_children().