testsuite: simplify sudo/sign-none tests in t2000-imp-exec.t
Problem: The IMP exec tests that use sign-none.toml and sudo are
unnecessarily complicated because the IMP has to be run in a subshell,
sometimes in the background. This not only duplicates a lot of code,
but makes it impossible to get the exit code of the IMP when it is
run in the background.

Add a function to run `flux-imp exec` under sudo with the
sign-none.toml config. While we're at it, create the `sleeper.sh`
helper at the beginning of the tests, instead of as part of another
test.

Update callers to use `imp_exec_sign_none()` where possible.