imp: exec: raise signal if child was terminated by one
Problem: The lingering IMP exits with 127+signo if the invoked child
was terminated by a signal during `flux-imp exec`, but ideally the
IMP would exit with the exact wait status of the child.

Add a convenience function, `imp_raise()`, which raises a signal
after setting the handler to the default disposition for the signal in
question.  Use imp_raise() to deliver the same signal which terminated
the child in `flux-imp exec` if the child is detected to have exited
due to a signal.