imp: cgroup: fix relative cgroup paths in containers
Problem: Relative paths for cgroups in /proc/self/cgroup may contain
leading `/..` when the IMP is run inside some containers. This is
because the path is relative to the container's mount point, not the
actual mount point of the cgroupfs.

Just strip the leading /.. from paths during discovery of the current
cgroup path from `/proc/self/cgroup`. This should only apply in
containers and thus CI.