imp: signals: do not use the SA_RESTART flag
Problem: The SA_RESTART flag is used when establishing signal handlers
in the IMP, but this is not necessary (affected system calls like
waitpid(2) already check for EINTR). It may also be convenient in
future code to allow blocking system calls to be interrupted by
signals to force the recheck of a condition, such as checking the
contents of cgroup.procs.

Drop the use of the SA_RESTART flag.