imp: exec: wait for empty cgroup before exiting
Problem: The IMP immediately exits after waitpid(2) returns for its
immediate child, but this could leave extra processes running in the
job's cgroup.

Wait for the job cgroup to be empty before allowing the IMP to exit.

An empty cgroup is considered one in which the only process is
cgroup.procs is the IMP itself.