testsuite: add test that ensures the IMP waits for an empty cgroup
Problem: No test ensures that imp exec waits for the job cgroup to
be empty before exiting.

Add a test that sends SIGTERM to an IMP managing a process tree.
This kills the immediate child of the IMP, and the IMP must wait
until SIGUSR1 is sent, which terminates the cgroup with SIGKILL.