imp: linger and forward signals to children of `flux-imp run`
Problem: Unlike `exec`, `flux-imp run` doesn't linger and instead
directly executes the target command. This then requires the use of
`flux-imp kill` to kill the privlieged process(es) that are a result.

Have the IMP fork() and exec() the run command. The lingering IMP
process can then forward signals delivered to it, including the use
of SIGUSR1 as a surrogate for SIGKILL.