imp: allow negative PID to be passed to imp_set_signal_child()
Problem: The current IMP signal interface does not allow a negative
PID to be passed to imp_set_signal_child(), and therefore signals can
only be forwarded to the direct child of the IMP. This works well with
`flux-imp exec`, where the child is a job shell which itself forwards
signals, but doesn't work as well in other situations where the IMP
is managing processes not running as a job.

Allow a negative PID less than -1 to be passed to
imp_set_signal_child().  This will allow the caller to request an
entire process group be signaled, instead of just a single child PID.