imp: run: put child in new process group
Problem: Before the removal of `flux-imp kill`, commands invoked by
`flux-imp run` were signaled by process group instead of an individual
process ID. This worked well because `flux-imp run` replaced itself
with the target command, and thus the command and its children were
in a process group (set up by flux-core's libsubprocess). Now, the
IMP itself will be in that process group, and if it tries to forward
signals to the entire process group it will end up re-signaling itself.

Call `setpgrp(2)` in the child to place it in its own process group.
The IMP now has a process group to which it can forward signals.