imp/cgroup: check correct fs in legacy path
Problem: In cgroup_init_path(), the legacy case should check
that the file system type of "/sys/fs/cgroup" is TMPFS, not the
file system type of "/sys/fs/cgroup/unified".

Correct this corner case by not re-using the local "fs" variable,
which was shared between multiple calls to statfs().  Use a new "fs"
variable for each call to statfs().