testsuite: fix t2000-imp-exec.t on some systems
Problem: A couple tests in t2000-imp-exec.t fail with

 write error: No such file or directory

when trying to write to `cgroup.procs` in a cgroup created for the
purposes of the test.

The issue appears to be that the test creates the cgroup at the
top level of the cgroup2 hierarchy, and a process can't be moved
from elsewhere in the hierarchy to a direct descendant of the root.
So the test fails if it is already running in a user slice or other
scope cgroup.

Get the cgroup path of the current process and use this as the parent
directory of the test cgroup. This seems to work around the issue.