docker: add --cgroupns=host to docker args
Problem: Tests in the flux-security testsuite that need to create a
temporary cgroup and add a process to it fail with the error "echo:
No such file or directory."

This appears to be occurring because moving a process from one cgroup
to another requires that the calling process have write permission
not only to cgroup.procs of the target cgroup, but also the common
ancestor of the target cgroup and the current cgroup.

Recent versions of docker mount the cgroupfs in the container in a
private namespace, so test cgroups are created in the private mount
root level cgroup directory (i.e. under `/`). This seems to confuse
the common ancestor check, and since the docker container will not
have write access to the root of the cgroup fs, the move of the
process fails (with the unhelpful ENOENT error).

The easiest fix for ci appears to be to add the option --cgroupns=host
to force the cgroup namespace in the container to match the host.
This allows tests to create new cgroups with the correct parent
directory, avoiding the error.